Elimina Polyline al finalizar ruta y muestra popup para añadir nombre y...

See merge request mayo.ortiz.adria/projecte4_entrebicis!13

+++ Backend/src/main/java/com/copernic/backend/Backend/entity/Sistema.java
@@ -48,7 +48,7 @@ public class Sistema {
      * Factor de conversió per transformar valors en saldo d'usuari.
      */
     @Column
-    private int conversioSaldo;
+    private double conversioSaldo;
 
     /**
      * Temps màxim permès per recollir una recompensa (en hores).
